+++ pycaption/sami.py
@@ -53,6 +53,7 @@ from .base import (
 )
 from .exceptions import (
     CaptionReadNoCaptions, CaptionReadSyntaxError, InvalidInputError,
+    CaptionReadTimingError
 )
 from .geometry import Layout, Alignment, Padding, Size
 
@@ -191,7 +192,11 @@ class SAMIReader(BaseReader):
         milliseconds = 0
 
         for p in sami_soup.select(f'p[lang|={language}]'):
-            milliseconds = int(float(p.parent['start']))
+            start_str = p.parent.get('start')
+            if not start_str:
+                raise CaptionReadTimingError(
+                    f"Missing start time on the following line: {p.parent}.")
+            milliseconds = int(float(start_str))
             start = milliseconds * 1000
             end = 0